refactor(Tag): create shared type
