+++ client/src/api/types/tag.ts
@@ -1,4 +1,4 @@
-import { TagType } from '../../types/tag-type'
+import { TagType } from '~shared/types/base'
 import { BaseModelParams } from './common'
 
 export type BaseTagDto = BaseModelParams & {
