+++ client/src/components/PostItem/PostItem.component.jsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { Link } from 'react-router-dom'
 import { useAuth } from '../../contexts/AuthContext'
-import { TagType } from '../../types/tag-type'
+import { TagType } from '~shared/types/base'
 import EditButton from '../EditButton/EditButton.component'
 import { RichTextFrontPreview } from '../RichText/RichTextEditor.component'
 import TagBadge from '../TagBadge/TagBadge.component'
