--- client/src/types/tag-type.ts
@@ -1,4 +1,0 @@
-export enum TagType {
-  Agency = 'AGENCY',
-  Topic = 'TOPIC',
-}
