--- server/src/types/tag-type.ts
@@ -1,4 +1,0 @@
-export const TagType = {
-  AGENCY: 'AGENCY',
-  TOPIC: 'TOPIC',
-}
