+++ shared/types/base/index.ts
@@ -1,4 +1,5 @@
 export * from './agency'
 export * from './answer'
 export * from './post'
+export * from './tag'
 export * from './user'
